src/tests/requirements.txt
pygame
numpy
websocket-client
PyOpenGL
PyOpenGL_accelerate

